Add files via upload
Detect a Cycle in a Linked List
In this article, we will solve the most asked interview question: Detect a Cycle in a Linked List
Problem Statement: Given head, the head of a linked list, determine if the linked list has a cycle in it. There is a cycle in a linked list if there is some node in the list that can be reached again by continuously following the next pointer.
Return true if there is a cycle in the linked list. Otherwise, return false.